debugging the plugin
