+++ docs/conversion_examples_gallery/recording/maxone.rst
@@ -11,10 +11,16 @@ Convert MaxOne data to NWB using :py:class:`~neuroconv.datainterfaces.ecephys.ma
 
 .. code-block:: python
 
+    >>> import os
     >>> from datetime import datetime
     >>> from dateutil import tz
     >>> from pathlib import Path
     >>> from neuroconv.datainterfaces import MaxOneRecordingInterface
+    >>>
+    >>> # If you do not have the Maxwell HDF5 plugin installed, call this first to automatically set it up.
+    >>> hdf5_plugin_path = f"{HDF5_PLUGIN_PATH}"
+    >>> MaxOneRecordingInterface.auto_install_maxwell_hdf5_compression_plugin(hdf5_plugin_path=hdf5_plugin_path)
+    >>> os.environ['HDF5_PLUGIN_PATH'] = hdf5_plugin_path
     >>>
     >>> file_path = f"{ECEPHY_DATA_PATH}/maxwell/MaxOne_data/Record/000011/data.raw.h5"
     >>> # Change the file_path to the location in your system
