+++ src/neuroconv/datainterfaces/ecephys/maxwell/maxonedatainterface.py
@@ -1,7 +1,10 @@
 """Authors: Cody Baker."""
-from ..baserecordingextractorinterface import BaseRecordingExtractorInterface
+from typing import Optional
 
-from ....utils.types import FilePathType
+from neo.rawio.maxwellrawio import auto_install_maxwell_hdf5_compression_plugin
+
+from ..baserecordingextractorinterface import BaseRecordingExtractorInterface
+from ....utils.types import FilePathType, FolderPathType
 
 
 class MaxOneRecordingInterface(BaseRecordingExtractorInterface):
@@ -12,6 +15,19 @@ class MaxOneRecordingInterface(BaseRecordingExtractorInterface):
     """
 
     ExtractorName = "MaxwellRecordingExtractor"
+
+    @staticmethod
+    def auto_install_maxwell_hdf5_compression_plugin(hdf5_plugin_path: Optional[FolderPathType] = None):
+        """
+        If you do not yet have the Maxwell compression plugin installed, this function will automatically install it.
+
+        Parameters
+        ----------
+        hdf5_plugin_path: string or Path, optional
+            Path to your systems HDF5 plugin library.
+            Default is None.
+        """
+        auto_install_maxwell_hdf5_compression_plugin(hdf5_plugin_path=hdf5_plugin_path)
 
     def __init__(self, file_path: FilePathType, verbose: bool = True):
         """
