+++ src/neuroconv/datainterfaces/ecephys/maxwell/maxtwodatainterface.py
@@ -1,8 +1,6 @@
 """Authors: Cody Baker."""
 from typing import List, Optional
 
-from neo.rawio.maxwellrawio import auto_install_maxwell_hdf5_compression_plugin
-
 from .maxonedatainterface import MaxOneRecordingInterface
 from ....utils.types import FilePathType, FolderPathType
 
@@ -17,19 +15,6 @@ class MaxTwoRecordingInterface(MaxOneRecordingInterface):
     """
 
     ExtractorName = "MaxwellRecordingExtractor"
-
-    @staticmethod
-    def auto_install_maxwell_hdf5_compression_plugin(hdf5_plugin_path: Optional[FolderPathType] = None):
-        """
-        If you do not yet have the Maxwell compression plugin installed, this function will automatically install it.
-
-        Parameters
-        ----------
-        hdf5_plugin_path: string or Path, optional
-            Path to your systems HDF5 plugin library.
-            Default is None.
-        """
-        auto_install_maxwell_hdf5_compression_plugin(hdf5_plugin_path=hdf5_plugin_path)
 
     @staticmethod
     def get_recording_names(file_path: FilePathType) -> Optional[List[str]]:
